fix(oxlint/lsp): run diagnostics after delayed worker initialization (#16244)
When the Client does not a configuration object in the `initialize` request, we delay the workspace worker start until we request later the configuration inside the `initialized` notification.

In the time, the client could already send `document/didOpen` notification, we should check here if we have already something in the file system.

- close https://github.com/oxc-project/oxc-zed/issues/31
- close https://github.com/oxc-project/oxc/issues/15374