fix(linter): Fix the radix rule docs to correctly render as a string-only config option (#16248)
And also reorganize/refactor the code to simplify it.

This fixes a mistake I made when initially setting up the docs generation for this rule.

Generated docs:

```md
## Configuration

This rule accepts one of the following string values:

### `"always"`

Always require the radix parameter when using `parseInt()`.

### `"as-needed"`

Only require the radix parameter when necessary.
```