refactor(linter): Update various rules to use DefaultRuleConfig pattern (#16249)
Basically just refactoring a bunch of rules to use the `DefaultRuleConfig` pattern, which conveniently also allows us to ensure that the serialization works as-expected with how the docs are rendering these config options.

The only real change here, other than to some diagnostics and a handful of docs, is the fix in `eslint/no-unused-expression` to ensure the docs for the `enforceForJSX` option capitalize it correctly, which I discovered because I updated to use this pattern and that caused the tests to - correctly - fail.

I ran the `just website` command and confirmed that the only docs changes from this branch vs main were the `no-unused-expression` fix and the minor docs update I did on no-bitwise.