fix(linter/plugins): reset state after error during AST visitation (#16246)
If there's an error during AST visitation, ensure state is reset to free memory.

This is only observable if user were to schedule some code in a later tick to access e.g. `sourceCode.text`, and that runs before the next file is linted. I've not added tests, as it'd be hard to write one which isn't flaky.