fix(linter): Fix config docs for no-return-assign and unicode-bom rules. (#16260)
These rules both take string options and were incorrectly documented as objects.

```
## Configuration

This rule accepts one of the following string values:

### `"always"`

Disallow all assignments in return statements.

### `"except-parens"`

Allow assignments in return statements only if they are enclosed in parentheses.
This is the default mode.
```

```md
## Configuration

This rule accepts one of the following string values:

### `"always"`

Always require a Unicode BOM (Byte Order Mark) at the beginning of the file.

### `"never"`

Never allow a Unicode BOM (Byte Order Mark) at the beginning of the file.
This is the default option.
```