feat(docs): Add links to the tsgolint source code for type-aware rules. (#16313)
I thought it was a bit weird to have the rule pages for type-aware rules link to the oxc rule definitions, when there's barely any relevant code in the linked "source" file. So I updated the rule docs to also link to tsgolint when relevant.

This doesn't support a specific git ref and is somewhat naively assuming that all rules in the tsgolint repo have a consistent path structure. However, the git ref can be added in the future if we need it, and this seems to work for all type-aware rules I've looked at so far.

Result:

<img width="758" height="260" alt="Screenshot 2025-11-30 at 7 01 30 PM" src="https://github.com/user-attachments/assets/eceb30e2-859b-481e-8faf-6e305474ad14" />