perf(linter/plugins): do not remove `messageId` field from `DiagnosticReport` before sending to Rust (#16442)
As @camc314 pointed out in https://github.com/oxc-project/oxc/pull/16296#pullrequestreview-3524629951:

It's probably cheaper to leave the `messageId` field present in the JSON sent over to Rust and let `serde` handle skipping it, than to loop through all the diagnostics on JS side to remove it.

`RuleTester` should be moved into a separate NPM package (#16018), at which point we can do a different build for that package, and not have `messageId` stored in `DiagnosticReport` in the first place in the main build. So then this issue will become moot.