perf(linter/plugins): move result-processing work off main JS thread (#16456)
Any work done in the callback to `call_with_return_value` (from my understanding) happens on main JS thread. JS plugins run single-threaded on JS side, so JS main thread is the main bottleneck for the whole linting process when JS plugins are in use.

Do only the minimal amount of work in these callbacks, and return the result as quickly as possible to the Rust thread, so that JS main thread is freed up to do more work.

In particular, don't deserialize JSON in the callbacks - release JS to continue linting the next file, and process the JSON afterwards on the calling thread.