refactor(linter/plugins): simplify return value of `load_plugin` (#16459)
Previously `load_plugin` returned a `Result<LoadPluginResult>`, where `LoadPluginResult` was an enum with one variant being an error case - so it was essentially `Result<Result<Data>>`.

Flatten it so that `LoadPluginResult` is a now struct containing details of success case, and the error case that `LoadPluginResult::Failure` previously covered becomes `Err`.

This simplifies the logic, and matches the pattern used by `lint_file` + `LintFileResult`.