refactor(linter/plugins): add debug assert for passing data from NAPI callback (#16465)
Alternative to #16463.

I think I've understood the situation now, and why Brooooklyn's original code ignored the result of `tx.send()`.

It is genuinely impossible for the call to `tx.send()` to fail, and that doesn't depend on there not being a bug in NAPI-RS - it's provable just from looking at the code in these functions, regardless of what NAPI-RS may do.

Add comments explaining why we can safely ignore the result of `tx.send()`, and add a `debug_assert!` just to make sure.

We want to absolutely minimize the overhead of moving between JS and Rust, so I think it's appropriate to leave out error handling for impossible cases like these.