fix(linter/plugins): improve TS types for tokens (#16502)
Improve the type def for `Token`.

Previously `Token` included `CommentToken`. This wasn't quite right because the `tokens` property of `Program` is `Token[]`, but it doesn't include comments.

So make `Token` the type for just code tokens, and use `Token | CommentToken` everywhere where the value can be either a code token or a comment token.

Also, remove the `BlockCommentToken` and `LineCommentToken` types. It seems unnecessarily detailed. `CommentToken` now covers both with `type: "Line" | "Block"`.