chore(linter): Update whitespace trim in a few rule files (#16553)
Otherwise the next person that edits these files will end up with goofy problems if they use an editor that actually follows the editorconfig (like mine does).

I found the problematic files by just doing a find+replace on "What it does", since that should be in every rule file, which caused my editor to save and applied to the editorconfig settings accordingly. The only one I left with this problem was `curly`, as it seemed to be reliant on the weird whitespace.

And then also replaces some usage of tabs in some rule tests with equivalent spaces. None of these rules are stylistic, so the difference is not relevant to the behavior of the rule.