fix(linter): Fix test in eqeqeq rule.
The test only worked because the fallback logic was bad previously. This config object was always wrong, and never should have worked.

This is, however, technically a breaking change. If any users were reliant on the behavior of the rule defaulting to `ignore` if you gave it a bad value, this fix may break things?