fix(linter): Fix jest/expect-expect rule docs. (#16550)
Fixes #16491.

The default value here is a bit weird so I had to adjust the docs to make it clear that there's a distinction with jest vs vitest.

Generated docs:

```md
## Configuration

This rule accepts a configuration object with the following properties:

### additionalTestBlockFunctions

type: `string[]`

default: `[]`

An array of function names that should also be treated as test blocks.

### assertFunctionNames

type: `string[]`

default: `["expect"]`

A list of function names that should be treated as assertion functions.

NOTE: The default value is `["expect"]` for Jest and
`["expect", "expectTypeOf", "assert", "assertType"]` for Vitest.

```