feat(oxfmt): Pass populated config to prettier formatting (#16584)
Fixes #16530

- Populate raw JSON config with parsed `FormatOptions`

If there is one concern, it is how to handle values currently supported only by `oxc_formatter`, such as `objectWrap: always`. 🤔

Now it will fall back to `preserve`, which produces the intended result rather than `collapse`.
However, this falls into the same category as the issue where `oxc_formatter` is not adapted for JS-in-XXX part.