refactor(linter): rename `GlobalValue::Writeable` to `Writable` (#16609)
"Writeable" is a mis-spelling. In addition:

Oxlint docs say:

> You may also use `"writeable"` or `true` to represent `"writable"`.

[ESLint's docs](https://eslint.org/docs/v8.x/use/configure/language-options#using-configuration-files-1) state "writable" (without an "e") as the name of the option.

This suggests that "writable" is the "real" name of this option, and "writeable" (with an "e") is an alias provided for backwards compat.

Therefore, I think it makes more sense to call the enum variant `GlobalValue::Writable` (with the correct spelling).

Main motivation is that it means globals get serialized with the right spelling, allowing removing a JS-side workaround from JS plugins code (#16610).

For consistency, I've changed "writeable" to "writable" in all test cases, except for 1 test which specifically aims to check that that "legacy" spelling is accepted.

https://github.com/oxc-project/oxc/blob/00c5614d1f1bb3465ba7aa5ad668a936b8ec1266/crates/oxc_linter/src/config/globals.rs#L161-L168