perf(formatter): optimize formatting of JSX element/fragment with a single child (#16631)
Add `FormatSingleChild` for optimizing JSX elements/fragments with a single meaningful child.

When there is only a single child, we do not need to write the child into a temporary buffer and then take it when formatting. Instead, we can directly write the child into the root buffer.

Also, this can avoid calling costly `best_fitting!` formatting in some situations.