refactor(ast): change capitalization of `CommentKind` variants (#16640)
Pure refactor. "multiline" is OK, but "singleline" is ungrammatical - it should be hyphenated ("single-line"). So rename the `CommentKind` variant to `SingleLineBlock`. Also rename `MultilineBlock` -> `MultiLineBlock` for consistency.

Obviously, this is not important. But as changing the variants of `CommentKind` is already a breaking change, it seems like the right moment to nitpick - and avoid getting stuck with it, or having to make another breaking change later.