test(linter/plugins): add conformance tests (#16650)
Add conformance tester for JS plugins.

The tester runs all of ESLint's test cases for all its built-in rules, but using Oxlint's `RuleTester` in place of ESLint's. i.e. the ESLint rules are run as Oxlint JS plugins.

It outputs the results in `snapshot.md` file.

Majority of the failing test cases are false positives. Later PRs will improve the conformance tester and `RuleTester` so it correctly handles `sourceType`, `jsx`, and TypeScript sources.

These tests do not run in CI. For now it's just useful for me to commit this, while I work on improving the tester.

I'm not sure if inside `apps/oxlint` is the right place for this ultimately, but for now I think it's passable.