test(linter/plugins): conformance tester error on test cases with custom parsers (#16716)
Some ESLint test cases provide a custom parser. We cannot easily support that, so just throw an error if a custom parser is provided.

Some of these cases were previously passing (Oxc managed to parse them), but some were not. It's hard when debugging to figure out for each test case whether it should be failing or not due to absence of custom parser, so simpler to just error for them all at this stage.