build(linter/plugins): add conformance build (#16751)
Supporting conformance testing is requiring adding extra code paths to `RuleTester` and main JS plugins code.

Previously, these paths were guarded by `if (DEBUG)` branches, so the extra code is not in release build. But ideally we don't want this code in debug build either, so the tests are testing a build that's identical to release build, except for enabling debug assertions.

Introduce a new build specifically for conformance testing, and gate code specific to conformance testing with `if (CONFORMANCE)` instead of `if (DEBUG)`.