feat(oxfmt): add `insert_final_newline` option
closes #15066

The difficult part here is only calculate line ending at the really end,
just before write, to avoid memory allocations on the source text when
you insert a newline.

@leaysgur This feature also requires a single file formatter so we can apply
this to node API and file system write. The single file formatter need
to accept a `Writer` so we can write to string or file.