test(linter/plugins): remove defunct test case (#16794)
This test case is defunct since #14401. It doesn't test what it's meant to.

From the comment in the plugin used in this test fixture:

```js
// Purpose of this test fixture is to ensure that source text and AST are available in `after` hook
// via `context.sourceCode` when the AST is not traversed
```

But the visitor returned by `createOnce` contains a visit function for `Program`, so the AST *is* traversed. On the other hand, if we remove that `Program` visit method, then it's an empty visitor, so the AST won't be traversed, but the `after` hook won't run either.

Therefore, the case which this fixture is designed to test is impossible. Delete the fixture.