test(linter/plugins): tests use Oxc parser (#16803)
Pure refactor of tests.

Unit tests for `isSpaceBetween` were using `@typescript-eslint/typescript-estree` to parse code snippets.

Instead use Oxc parser, and set up global state the same way as if the file was being linted, instead of mocking `source_code.ts` to get the source text into the linter's internals.

This brings the tests closer to real usage. It also prepares the way for changing how tokens are parsed in a following PR (#16805).