refactor(linter): Update various rules to use DefaultRuleConfig and fix some minor issues. (#16816)
Update a handful of rules to use the DefaultRuleConfig pattern.

- `react/button-has-type`
- `react/jsx-no-target-blank`
- `react/style-prop-object`
- `vue/define-props-destructuring`

For `react/style-prop-object`, this _was_ sorting the `allow` vector to enable binary search, but I'd be surprised if the performance of this matters enough to warrant it. And it wasn't being done consistently anyway, the most common case (JSX elements) wasn't even using a binary search.

This also fixes the rule docs for `vue/define-props-destructuring`, which were previously broken.

Now they look like this:

```md
## Configuration

This rule accepts a configuration object with the following properties:

### destructure

type: `"always" | "never"`

default: `"always"`

Require or prohibit destructuring.

#### `"always"`

Requires destructuring when using `defineProps` and warns against using `withDefaults` with destructuring

#### `"never"`

Requires using a variable to store props and prohibits destructuring
```