feat(eslint-plugin-jest): Add fix capabilities to prefer to contain rule (#16809)
Currently oxclint can't fix any `prefer-to-contain` error, but the [jest eslint rule can](https://github.com/jest-community/eslint-plugin-jest/blob/main/docs/rules/prefer-to-contain.md), also this will be helpful to vitest.

Any alternative for the trailing comma check are welcome. None of my global searched in the codebase lead to an example, only to check if a trailing comma exists.

---------

Co-authored-by: Cameron Clark <cameron.clark@hey.com>