refactor(linter): Simplify from_configuration for several tsgolint rules. (#16851)
This updates all the tsgolint rules that were using a box wrapper + `DefaultRuleConfig<RuleNameConfig>`. And also one react rule.

The default_true utility is also unneeded, as the defaults will work fine without it. The `impl Default` sets the defaults for the relevant fields on each of these rules.

Basically the same kind of change as #16818.