chore(linter/plugins): Add more values to globals fixture test. (#16860)
Add a few other values in the oxlintrc.json file to ensure that the
various possible input values are tested. These should work and be
parsed correctly by the config deserializer, generally.

The "weird" values are booleans and readable/writeable. They are
supported for backwards-compatibility with older ESLint configs.

cc: @overlookmotel not sure if this is desired, or if this should be a
separate test file? Just making sure we have support for the various
globals variants that need to be accepted.