refactor(linter): check valid Vitest calls with pattern matching (#16924)
This trades compile time + binary size for maybe slightly less performance when checking Vitest function calls. We currently have to manually maintain this list which is a bit difficult to do as aliases make this list grow quickly. This is not sustainable long-term if more functions and aliases get added, so I'm proposing we make the switch now.

As part of this I also removed support for `fdescribe`, `fit`, `xdescribe`, `xit`, etc. as these don't actually seem to be supported in Vitest? At least I didn't find any references to them in the docs or in the source code, so I think this might have been a mistake when copying from the Jest version.