fix(linter): Prevent extra fields from being present on oxlint config file (#16874)
This adds a `$schema` field to the JSON Schema, but otherwise disallows any fields not explicitly defined in the Oxlintrc struct.

This ensures that we can warn the user if they have invalid config fields:

<img width="695" height="417" alt="Screenshot 2025-12-14 at 11 22 02 PM" src="https://github.com/user-attachments/assets/8dc2252e-58ad-4bb5-a84e-cd468ad35941" />

The improved enforcement ensures the correct usage of the configuration file, and this change found a bug in another rule (see #16873).