feat(linter/eslint-plugin-vitest): implement prefer-spy-on (#16426)
Related to #4656 

Implement
[prefer-spy-on](https://github.com/vitest-dev/eslint-plugin-vitest/blob/main/docs/rules/prefer-spy-on.md)
lint ruler.

~~It's my first contribution made in Rust, and in Oxc, any suggestion to
improve the code it's welcome. The logic is based on the JS rule
implementation.~~

~~Update 12/04/2025: I have just notice Jest have a similar linter rule,
so I have duplicate the code and adapt it to vitest. I discarded the
option to abstract all the logic as none can ensure that both lints rule
are going to be the same in the future. Discarded the map between jest
and vitest rules as the documentacion for this specific rule is
distinct.~~

Update 12/15/2025: The rule is implemented as a compatible with jest
instead of duplicate code:

Oxlint migrate PR:
https://github.com/oxc-project/oxlint-migrate/pull/291