fix(linter/plugins): apply defaults from `meta.schema` to options (#16930)
ESLint has 2 different ways to define default options for a rule - `rule.meta.defaultOptions`, or `rule.meta.schema` containing a schema which defines default values.

We previously supported the former, but not the latter. This PR implements support for defining defaults via `rule.meta.schema`.

We compile the schema and apply it to options on JS side, using [`ajv`](https://www.npmjs.com/package/ajv), which is what ESLint uses. But this process only happens once before the start of linting, not once for every file, so the perf hit should be small.

This also prepares the way for validating options against the schema (#15631).