feat(linter): Add ignoreCase option to `react/jsx-no-duplicate-props` rule.
This adds the ignoreCase option to the rule. I am unsure if the move from Atom to Cow is necessary, but I couldn't figure out how to get it converted to lowercase as an Atom without converting it to something else anyway.

Adds tests from the upstream rule as well.

Fixes #16968.