feat(linter/eslint-plugin-jest): Add fix capabilities to prefer-called-with rule (#16987)
Jest and Vitest plugin have a diff in implementation, while the first
one lack of a fix the second have it. Both share the same logic
detection. For the sake of reutilization and parity functionality, the
jest rule now can fix the error codes.

A following PR will add this rule as vitest compatible, I'm not adding
all together so the PRs only add one new thing.