fix(linter): Fix `vitest/no-restricted-vi-methods` and add tests for it. (#16971)
This fixes a few things:

- Fixes the config builder so `vitest/no-restricted-vi-methods` can actually be enabled via the config file.
- Fixes disabling the `vitest/no-restricted-vi-methods` rule via a disable directive comment.

`vitest/no-restricted-vi-methods` wasn't being remapped correctly by the config builder, as the config builder only remapped `jest/any-name` to `vitest/any-name`, and there was no special handling for rules which need to map to an entirely different name. This has added that logic.

`vitest/no-restricted-jest-methods` _does_ work, but that's not what it should be called or how it should work.

Also a minor docs improvements.

I'm unsure if this is sufficiently performant, honestly, but it fixes the problem. There is very likely a better way to do this.

As an aside, the disable comment logic seems... very prone to failure if we're just doing a `name.contains(rule_name)` in most cases. That'd allow `// oxlint-disable-next-line a` to work on any rule with an `a` in the name, I think?