feat(linter/eslint-plugin-vitest): Add prefer-called-with as vitest compatible jest rule (#16993)
Related to https://github.com/oxc-project/oxc/issues/4656

# Summary 
Add `prefer-called-with` as a vitest compatible rule. 

Oxlint migrate PR:
https://github.com/oxc-project/oxlint-migrate/pull/295

# Comment
This is the last PR of compatible vitest rule with the current codebase.
So the following rules are going to be implement from zero, so the
cadency of PR will be lower. The good news most of the remaining rules
also apply to jest, so at the end bot plugin will have the same parity.