perf(ast): `#[ast]` macro use `#[repr(transparent)]` for single-field structs (#17052)
`#[ast]` macro adds `#[repr(C)]` to structs. For structs with a single field, use `#[repr(transparent)]` instead. This does not alter memory layout, but can change the ABI, which can make it more performant to pass such structs to functions in some cases, as it can change the registers used for passing.

This will help with making lexer tokens serializable (#17025), since `Token` will have `#[ast]` added to it, and we need it to continue to be `#[repr(transparent)]`.