refactor(formatter): move the parsing group for sorting imports to the place where it is used (#17064)
Only parse the groups when needed, as I noticed it would return early when the document only has a hard line (which means the file is empty). Then parsing the group is a waste.

Additionally, I realized we may improve this further by moving this part into the CLI, since the `experimental_sort_imports` option is always the same in each file, but currently, we have to parse again and again in each file